+++ source/renderer/app/themes/daedalus/dark-blue.js
@@ -340,7 +340,13 @@ export default {
   '--theme-staking-content-background-color': '#263345',
   '--theme-staking-content-border-color': '#263345',
   '--theme-staking-font-color-accent': '#cecfd1',
-  '--theme-staking-font-color-regular': '#0b1926',
+  '--theme-staking-font-color-regular': 'rgba(233, 244, 254, 1)',
+  '--theme-staking-font-color-light': 'rgba(233, 244, 254, 0.7)',
+  '--theme-staking-font-color-lighter': 'rgba(233, 244, 254, 0.5)',
+  '--theme-staking-table-head-background-color': '#536370',
+  '--theme-staking-table-border-color': '#c6cdd6',
+  '--theme-staking-link-color': 'rgba(32, 181, 107, 1)',
+  '--theme-staking-link-color-light': 'rgba(32, 181, 107, 0.5)',
 
   '--theme-switch-label-color': '#e9f4fe',